bot/requirements.txt
pylint>=1.7.4
python-telegram-bot>=12.0.0b1
requests>=2.21.0